+++ src/fight.c
@@ -460,8 +460,8 @@ void one_hit( CHAR_DATA *ch, CHAR_DATA *victim, int dt )
     }
     else
     {
-	thac0_00 = class_table[ch->class].thac0_00;
-	thac0_32 = class_table[ch->class].thac0_32;
+	thac0_00 = pc_race_table[ch->race].thac0_00;
+	thac0_32 = pc_race_table[ch->race].thac0_32;
     }
     thac0  = interpolate( ch->level, thac0_00, thac0_32 );