+++ src/magic.c
@@ -234,7 +234,7 @@ bool saves_spell( int level, CHAR_DATA *victim, int dam_type )
 	case IS_VULNERABLE:	save -= 2;	break;
     }
 
-    if (!IS_NPC(victim) && class_table[victim->class].fMana)
+    if (!IS_NPC(victim) && pc_race_table[victim->race].fMana)
 	save = 9 * save / 10;
     save = URANGE( 5, save, 95 );
     return number_percent( ) < save;
@@ -549,7 +549,7 @@ void do_cast( CHAR_DATA *ch, char *argument )
     else
     {
         ch->mana -= mana;
-        if (IS_NPC(ch) || class_table[ch->class].fMana) 
+        if (IS_NPC(ch) || pc_race_table[ch->race].fMana) 
 	/* class has spells */
             (*skill_table[sn].spell_fun) ( sn, ch->level, ch, vo,target);
         else