+++ src/update.c
@@ -80,11 +80,11 @@ void advance_level( CHAR_DATA *ch, bool hide )
     set_title( ch, buf );
 
     add_hp	= con_app[get_curr_stat(ch,STAT_CON)].hitp + number_range(
-		    class_table[ch->class].hp_min,
-		    class_table[ch->class].hp_max );
+		    pc_race_table[ch->race].hp_min,
+		    pc_race_table[ch->race].hp_max );
     add_mana 	= number_range(2,(2*get_curr_stat(ch,STAT_INT)
 				  + get_curr_stat(ch,STAT_WIS))/5);
-    if (!class_table[ch->class].fMana)
+    if (!pc_race_table[ch->race].fMana)
 	add_mana /= 2;
     add_move	= number_range( 1, (get_curr_stat(ch,STAT_CON)
 				  + get_curr_stat(ch,STAT_DEX))/6 );
@@ -177,7 +177,7 @@ int hit_gain( CHAR_DATA *ch )
     else
     {
 	gain = UMAX(3,get_curr_stat(ch,STAT_CON) - 3 + ch->level/2); 
-	gain += class_table[ch->class].hp_max - 10;
+	gain += pc_race_table[ch->race].hp_max - 10;
  	number = number_percent();
 	if (number < get_skill(ch,gsn_fast_healing))
 	{
@@ -251,7 +251,7 @@ int mana_gain( CHAR_DATA *ch )
 	    if (ch->mana < ch->max_mana)
 	        check_improve(ch,gsn_meditation,TRUE,8);
 	}
-	if (!class_table[ch->class].fMana)
+	if (!pc_race_table[ch->race].fMana)
 	    gain /= 2;
 
 	switch ( ch->position )